added new logic for landing
