+++ app-landing/data/api/tokens/index.js
@@ -1,7 +1,3 @@
-import tokensClaimErc20 from './tokens-claim-erc-20'
-import tokensClaimErc721 from './tokens-claim-erc721'
-
 export {
-  tokensClaimErc20,
-  tokensClaimErc721
+
 }
