+++ app-landing/data/store/index.js
@@ -4,7 +4,7 @@ import thunk from 'redux-thunk'
 import { connectRouter, routerMiddleware } from 'connected-react-router'
 import saga from './saga'
 import { createBrowserHistory } from 'history'
-import { user, tokens } from './reducers'
+import { user } from './reducers'
 const sagaMiddleware = createSagaMiddleware()
 export const history = createBrowserHistory()
 
@@ -12,7 +12,6 @@ export default () => {
   const store = createStore(
     combineReducers({
       user,
-      tokens,
       router: connectRouter(history)
     }),
     {},
