+++ app-landing/data/store/reducers/index.js
@@ -1,7 +1,5 @@
 import user from './user'
-import tokens from './tokens'
 
 export {
-  user,
-  tokens
+  user
 }
