--- app-landing/data/store/reducers/tokens/reducers/index.js
@@ -1,7 +1,0 @@
-import setLink from './set-link'
-import setLoading from './set-loading'
-
-export default {
-  setLink,
-  setLoading
-}
