--- app-landing/data/store/reducers/tokens/reducers/set-amount.js
@@ -1,1 +1,0 @@
-export default (state, { payload: { amount } }) => ({ ...state, amount })
