--- app-landing/data/store/reducers/tokens/reducers/set-link.js
@@ -1,1 +1,0 @@
-export default (state, { payload: { link } }) => ({ ...state, link })
