--- app-landing/data/store/reducers/tokens/reducers/set-loading.js
@@ -1,1 +1,0 @@
-export default (state, { payload: { loading } }) => ({ ...state, loading })
