+++ app-landing/data/store/reducers/user/reducers/index.js
@@ -2,7 +2,6 @@ import changeLocale from './change-locale'
 import setWallet from './set-wallet'
 import setStep from './set-step.js'
 import setLoading from './set-loading'
-import setTransactionId from './set-transaction-id'
 import setErrors from './set-errors'
 
 export default {
@@ -10,6 +9,5 @@ export default {
   setWallet,
   setStep,
   setLoading,
-  setTransactionId,
   setErrors
 }
