--- app-landing/data/store/reducers/user/reducers/set-transaction-id.js
@@ -1,1 +1,0 @@
-export default (state, { payload: { transactionId } }) => ({ ...state, transactionId })
