+++ app-landing/data/store/saga/index.js
@@ -1,7 +1,7 @@
-import tokens from './tokens'
+import user from './user'
 
 function * saga () {
-  yield * tokens()
+  yield * user()
 }
 
 export default saga
