--- app-landing/data/store/saga/tokens/every/claim-tokens.js
@@ -1,27 +1,0 @@
-import { put } from 'redux-saga/effects'
-import { delay } from 'redux-saga'
-
-const generator = function * ({ payload }) {
-  try {
-    yield put({ type: 'USER.SET_LOADING', payload: { loading: true } })
-    yield delay(2000)
-    yield put({ type: 'USER.SET_TRANSACTION_ID', payload: { transactionId: '0xd003b90ae43dbd84691279a6c1ab93adb7e15299e11c619674acc267193d6ebc' } })
-    yield put({ type: 'USER.SET_STEP', payload: { step: 4 } })
-    yield delay(2000)
-    // yield put({
-    //   type: 'USER.SET_ERRORS',
-    //   payload: {
-    //     errors: ['LINK_EXPIRED'] // LINK_CANCELED, LINK_FAILED
-    //   }
-    // })
-    yield put({ type: 'USER.SET_STEP', payload: { step: 5 } })
-  } catch (e) {
-    console.error(e)
-  }
-}
-
-export default generator
-
-generator.selectors = {
-  wallet: ({ user: { wallet } }) => wallet
-}
