--- app-landing/data/store/saga/user/every/sign-in.js
@@ -1,8 +1,0 @@
-const generator = function * () {
-  try {
-  } catch (e) {
-    console.error(e)
-  }
-}
-
-export default generator
