+++ app-landing/data/store/saga/user/helpers/define-network-name.js
@@ -1,0 +1,7 @@
+export default ({ networkId }) => NETWORK_IDS[`_${networkId}`]
+
+const NETWORK_IDS = {
+  _1: 'homestead',
+  _3: 'ropsten',
+  _4: 'rinkeby'
+}
