+++ app-landing/data/store/saga/user/index.js
@@ -1,7 +1,7 @@
 import { takeEvery } from 'redux-saga/effects'
 
-import signIn from './every/sign-in'
+import checkBalance from './every/sign-in'
 
 export default function * () {
-  yield takeEvery('*USER.SIGN_IN', signIn)
+  yield takeEvery('*USER.CHECK_BALANCE', checkBalance)
 }
